setup: webpack.config.js needs to be at root of the project
